implementacion de CONTEXT
