+++ src/components/TodoCreateWeeklyPin/TodoCreateWeeklypin.css
@@ -2,9 +2,6 @@ label {
   padding-left: 10px;
   font-weight: 600;
 }
-.pin-data {
-  margin-bottom: 30px;
-}
 .pin-data input {
   width: 100%;
   height: 30px;
