+++ src/components/TodoWeeklyPin/TodoWeeklyPin.css
@@ -16,7 +16,7 @@
   display: flex;
   align-items: center;
   padding: 5px;
-  margin: 20px 0;
+  margin: 10px 0;
   margin-right: 10px;
   background-color: var(--main-color--);
   border-radius: 10px;
