+++ src/layouts/sectionLeft/SectionLeft.css
@@ -6,7 +6,7 @@
 .section-dates-header {
   display: flex;
   align-items: center;
-  margin-bottom: 50px;
+  margin-bottom: 30px;
 }
 .section-dates .header-logo {
   width: 25px;
@@ -28,7 +28,7 @@
   display: flex;
   flex-direction: column;
   justify-content: space-between;
-  height: 80vh;
+  height: 75vh;
 }
 .reminders-title button {
   border: none;
@@ -47,7 +47,8 @@
 .react-calendar {
   max-width: 100%;
   background: white;
-  border-radius: 10%;
+  border-radius: 10px;
+  padding: 10px;
 }
 
 .react-calendar button {
